Fix: Arweave verifier
